+++ HISTORY.md
@@ -1,4 +1,7 @@
 # Changelog
+# 0.6.0
+* ADD mzTAB support
+* ENH color Volcano Plot data points using list of protein names `color_list=your_protein_list`
 
 # 0.5.4
 * FIX altair version - binning of streamlit version